Local assets, code highlighting, better color management. All assets are now loaded locally (jQuery, font-awesome). They are kept in the `public` folder in each file's respective subdirectory.
Code highlighting is handled by Rouge (a Pygments compatible markdown codeblock renderer). Definitions for the highlights are defined in `_pygments.scss`.

All colors are now managed in `_variables.scss`. This includes shading and special definitions. For example, the colors for the navigation bar are defined at the end of `_variables.scss` with the prefix `$nav-`. This should help keep colors consistent across the entire site.