8pt grid system, consistent spacing across the board. This commit introduces the ideas of the 8pt grid system (see https://builttoadapt.io/intro-to-the-8-point-grid-system-d2573cde8632) and exposes utilities for using those sizes ($half-u, $u1, $u4, etc. The names c(/sh)ould be better).
All instances of element spacing (except inline with text, such as inline `code`) now use this 8pt grid, resulting in a very consistent, comfortable layout that can also be adjusted easily via just 1 variable ($u1).

Additionally, this commit adds an `80ch` width limit on topic samples. This way, the samples section is guaranteed to be 80 characters in width, and samples can be written accordingly with the knowledge that text fewer than 80 characters wide will always be displayed in one line.